+++ Активности/Активности/Dictionary1.xaml
@@ -70,55 +70,6 @@
             </Setter.Value>
         </Setter>
     </Style>
-    <Style x:Key="{x:Type local:PageButton}" TargetType="{x:Type local:PageButton}" BasedOn="{x:Null}">
-        <Setter Property="Foreground" Value="{DynamicResource WindowText}"/>
-        <Setter Property="HorizontalContentAlignment" Value="Center"/>
-        <Setter Property="VerticalContentAlignment" Value="Center"/>
-        <Setter Property="Padding" Value="0,-15,0,0" />
-        <Setter Property="Height" Value="25"/>
-        <Setter Property="Width" Value="25"/>
-        <Setter Property="FontFamily" Value="Segoe Print"/>
-        <Setter Property="FontSize" Value="24"/>
-        <Setter Property="FontWeight" Value="Bold"/>
-        <Setter Property="BorderThickness" Value="0"/>
-        <Setter Property="Foreground" Value="Black"/>
-        <Setter Property="Template">
-            <Setter.Value>
-                <ControlTemplate TargetType="{x:Type local:PageButton}">
-                    <!-- We use Grid as a root because it is easy to add more elements to customize the button -->
-                    <Grid x:Name="Grid">
-                        <Border x:Name="Border"
-                           BorderBrush="Transparent"
-                           BorderThickness="0"/>
-                        <!-- Content Presenter is where the text content etc is placed by the control -->
-                        <!-- The bindings are useful so that the control can be parameterized without editing the template -->
-                        <ContentPresenter HorizontalAlignment="{TemplateBinding HorizontalContentAlignment}"
-                                    Margin="{TemplateBinding Padding}"
-                                    VerticalAlignment="{TemplateBinding VerticalContentAlignment}"
-                                    RecognizesAccessKey="True"/>
-                    </Grid>
-                    <!--Each state sets a brush on the Border in the template -->
-                    <ControlTemplate.Triggers>
-                        <Trigger Property="IsKeyboardFocused" Value="true">
-                            <Setter Property="BorderBrush" Value="{DynamicResource DefaultedBorderBrush}" TargetName="Border"/>
-                        </Trigger>
-                        <Trigger Property="IsMouseOver" Value="true">
-                            <Setter Property="Background" Value="Transparent" TargetName="Border"/>
-                        </Trigger>
-                        <Trigger Property="IsPressed" Value="true">
-                            <Setter Property="Background" Value="{DynamicResource PressedBrush}" TargetName="Border"/>
-                            <Setter Property="BorderBrush" Value="{DynamicResource PressedBorderBrush}" TargetName="Border"/>
-                        </Trigger>
-                        <Trigger Property="IsEnabled" Value="true"/>
-                        <Trigger Property="IsEnabled" Value="false">
-                            <Setter Property="Background" Value="{DynamicResource DisabledBackgroundBrush}" TargetName="Border"/>
-                            <Setter Property="BorderBrush" Value="{DynamicResource DisabledBorderBrush}" TargetName="Border"/>
-                            <Setter Property="Foreground" Value="{DynamicResource DisabledForegroundBrush}"/>
-                        </Trigger>
-                    </ControlTemplate.Triggers>
-                </ControlTemplate>
-            </Setter.Value>
-        </Setter>
-    </Style>
+  
     <SolidColorBrush x:Key="ButtonBorderHover" Color="Transparent"/>
 </ResourceDictionary>
