--- Активности/Активности/PageButton.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Controls;
-
-namespace Активности
-{
-    class PageButton : Button
-    {
-    }
-}
