+++ Активности/Активности/PageButton.xaml
@@ -1,0 +1,33 @@
+﻿<UserControl x:Class="Активности.PageButton"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Активности"
+             mc:Ignorable="d" 
+             d:DesignHeight="300" d:DesignWidth="300">
+    <Grid>
+        <Button WindowChrome.IsHitTestVisibleInChrome="True" BorderThickness="0" Padding="0" Width="30" Height="30"
+                                                Background="{x:Null}" HorizontalAlignment="Left" VerticalContentAlignment="Stretch" VerticalAlignment="Top" HorizontalContentAlignment="Stretch">
+            <Button.Content>
+                <Grid>
+                    <Label Content="{Binding ContentPageButton}" x:Name="textPage" HorizontalContentAlignment="Center" VerticalContentAlignment="Center" FontSize="24"/>
+                    <Border CornerRadius="50" BorderThickness="2" Background="{x:Null}" BorderBrush="Black">
+                        <Border.Style>
+                            <Style TargetType="{x:Type Border}">
+                                <Style.Triggers>
+                                    <Trigger Property="IsMouseOver" Value="False">
+                                        <Setter  Property="Background" Value="Azure" />
+                                    </Trigger>
+                                    <Trigger Property="IsMouseOver" Value="True">
+                                        <Setter Property="Background" Value="Red" />
+                                    </Trigger>
+                                </Style.Triggers>
+                            </Style>
+                        </Border.Style>
+                    </Border>
+                </Grid>
+            </Button.Content>
+        </Button>
+    </Grid>
+</UserControl>
