+++ Активности/Активности/Pages/minimum.xaml.cs
@@ -25,89 +25,25 @@ namespace Активности.Pages
         public minimum()
         {
             InitializeComponent();
-            guild_counter();
-            build_counter();
-            pirate_counter();
         }
-        private void guild_x10_pluse_Click(object sender, RoutedEventArgs e)
-        {
-            guild_10.Value++;
-            guild_counter();
-        }
-
-        private void guild_x10_minuse_Click(object sender, RoutedEventArgs e)
-        {
-            guild_10.Value--;
-            guild_counter();
-        }
-        private void guild_build_pluse_Click(object sender, RoutedEventArgs e)
-        {
-            guild_build.Value++;
-            build_counter();
-        }
-
-        private void guild_build_minuse_Click(object sender, RoutedEventArgs e)
-        {
-            guild_build.Value--;
-            build_counter();
-        }
-        private void pirate_island_pluse_Click(object sender, RoutedEventArgs e)
-        {
-            pirate_island.Value++;
-            pirate_counter();
-        }
-
-        private void pirate_island_minuse_Click(object sender, RoutedEventArgs e)
-        {
-            pirate_island.Value--;
-            pirate_counter();
-        }
-        private void guild_counter()
-        {
-            string guild_10_count = Convert.ToString(guild_10.Value) + " из " + Convert.ToString(guild_10.Maximum);
-            guild_10_counter.Dispatcher.BeginInvoke(new Action(delegate ()
-            {
-                guild_10_counter.Content = guild_10_count;
-            }));
-        }
-
-        private void build_counter()
-        {
-            string guild_build_count = Convert.ToString(guild_build.Value) + " из " + Convert.ToString(guild_build.Maximum);
-            guild_build_counter.Dispatcher.BeginInvoke(new Action(delegate ()
-            {
-                guild_build_counter.Content = guild_build_count;
-            }));
-        }
-
-        private void pirate_counter()
-        {
-            string pirate_island_count = Convert.ToString(pirate_island.Value) + " из " + Convert.ToString(pirate_island.Maximum);
-            pirate_island_counter.Dispatcher.BeginInvoke(new Action(delegate ()
-            {
-                pirate_island_counter.Content = pirate_island_count;
-            }));
-        }
-        private void ContentControl_Click(object sender, RoutedEventArgs e)
+        private void NextPage_Click(object sender, RoutedEventArgs e)
         {
             var btn = sender as System.Windows.Controls.Button;
             string name = btn.Name;
             switch (name)
             {
-                case "necessarily":
-                    get{ return (Convert)GetValue(ContentBlock); }
-                    set{ SetValue(Content, new desirable()); }
+                case "page_1":
+                    ContentControlMinimum.Content = new minimum();
+                  //  ActivePage1.Visibility Property = false;
                     break;
-                case "additional":
-                    ContentBlock.Content = new additional();
+                case "page_2":
+                    ContentControlMinimum.Content = new page2Minimum();
                     break;
-                case "desirable":
-                    ContentBlock.Content = new desirable();
+                case "page_3":
+                    ContentControlMinimum.Content = new page3Minimum();
                     break;
-                case "minimum":
-                    ContentBlock.Content = new minimum();
-                    break;
-                default:
+                case "page_4":
+                    ContentControlMinimum.Content = new page4Minimum();
                     break;
             }
         }
